+++ core/network/src/main/java/com/wap/wapp/core/network/model/survey/form/SurveyQuestionResponse.kt
@@ -1,6 +1,8 @@
-package com.wap.wapp.core.network.model.survey
+package com.wap.wapp.core.network.model.survey.form
 
 import com.wap.wapp.core.model.survey.SurveyQuestion
+import com.wap.wapp.core.network.model.survey.QuestionTypeResponse
+import com.wap.wapp.core.network.model.survey.toDomain
 
 data class SurveyQuestionResponse(
     val questionTitle: String,
